refactor: create cli for programatically usage
