+++ lib/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   ast: require("./ast"),
   chakra: require("./chakra"),
+  cli: require("./cli"),
   utils: require("./utils"),
 };
